% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_estimators.R
\name{alpha_lad}
\alias{alpha_lad}
\title{Least Absolute Deviations Scaling Estimator}
\usage{
alpha_lad(x, yhat)
}
\arguments{
\item{x}{Numeric vector of observed values.}

\item{yhat}{Numeric vector of predicted values.}
}
\value{
Numeric scalar representing the estimated scaling parameter.
}
\description{
Computes a robust scaling parameter using the weighted median
of the ratios x / yhat.
}
\details{
This estimator minimizes the absolute error:

sum(|x - alpha * yhat|)

It is more robust to outliers than least squares.
}
\examples{
x <- rnorm(100)
yhat <- rnorm(100)
alpha_lad(x, yhat)

}
