% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/butterworth.R
\name{butterworth_polos_MZT}
\alias{butterworth_polos_MZT}
\title{Butterworth Filter to Equivalent AR Poles (MZT)}
\usage{
butterworth_polos_MZT(A, N, fs)
}
\arguments{
\item{A}{Stopband attenuation in dB.}

\item{N}{Filter order (positive integer).}

\item{fs}{Sampling frequency.}
}
\value{
A list containing filter parameters and poles.
}
\description{
Converts an analog Butterworth low-pass filter into
discrete-time poles using the Matched Z-Transform (MZT).
}
\details{
These poles define the equivalent autoregressive (AR)
representation used in the BTWAR framework.
}
\examples{
butterworth_polos_MZT(A = 40, N = 4, fs = 2)

}
