% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral.R
\name{compute_spectrum}
\alias{compute_spectrum}
\title{Compute Power Spectrum}
\usage{
compute_spectrum(x, fs)
}
\arguments{
\item{x}{Time series.}

\item{fs}{Sampling frequency.}
}
\value{
List with frequency vector and power spectrum.
}
\description{
Computes the one-sided power spectrum using the Fast Fourier Transform.
}
