% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_validation.R
\name{cv_butterworth_AsN}
\alias{cv_butterworth_AsN}
\title{Rolling-Origin Cross-Validation for BTWAR}
\usage{
cv_butterworth_AsN(
  y,
  As_vec,
  N_vec,
  fs,
  min_train = NULL,
  verbose = TRUE,
  method
)
}
\arguments{
\item{y}{Stationary time series.}

\item{As_vec}{Candidate attenuation values.}

\item{N_vec}{Candidate filter orders.}

\item{fs}{Sampling frequency.}

\item{min_train}{Minimum training size.}

\item{verbose}{Logical flag.}

\item{method}{Scaling estimation method.}
}
\value{
List containing CV results and optimal parameters.
}
\description{
Selects optimal Butterworth parameters (N, A)
using out-of-sample RMSE.
}
