% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btwar_fit.R
\name{fitted.btwar}
\alias{fitted.btwar}
\title{Fitted Values from a BTWAR Model}
\usage{
\method{fitted}{btwar}(object, ...)
}
\arguments{
\item{object}{Object of class \code{"btwar"}.}

\item{...}{Additional arguments (currently unused).}
}
\value{
Numeric vector of fitted values on the training set.
}
\description{
Fitted Values from a BTWAR Model
}
