% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{make_AR_dataset}
\alias{make_AR_dataset}
\title{Construct AR Supervised Dataset}
\usage{
make_AR_dataset(series, p = 5, train_frac = 0.7, scale = TRUE)
}
\arguments{
\item{series}{Numeric time series.}

\item{p}{AR order.}

\item{train_frac}{Training fraction (0-1).}

\item{scale}{Logical; whether to standardize series.}
}
\value{
List containing training/test splits and scaling info.
}
\description{
Converts a univariate time series into a supervised learning
dataset suitable for autoregressive modeling.
}
