% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stationarity.R
\name{make_stationary}
\alias{make_stationary}
\title{Make Time Series Stationary}
\usage{
make_stationary(x, max_d = 3, alpha = 0.05)
}
\arguments{
\item{x}{Numeric time series.}

\item{max_d}{Maximum number of differences allowed. Default is 3.}

\item{alpha}{Significance level for ADF test. Default is 0.05.}
}
\value{
A list containing:
\item{x_stationary}{Stationary transformed series}
\item{d}{Number of differences applied}
\item{p_value}{Final ADF test p-value}
\item{stationary}{Logical indicator}
}
\description{
Applies successive differencing until stationarity is achieved
according to the Augmented Dickey-Fuller (ADF) test.
}
\details{
Differencing stops when either:
\itemize{
\item The ADF test rejects the unit root hypothesis, or
\item The maximum number of differences is reached.
}
}
\examples{
set.seed(1)
x <- cumsum(rnorm(200))
make_stationary(x)

}
