% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ar_tools.R
\name{polos_AR}
\alias{polos_AR}
\title{Compute AR Model Poles}
\usage{
polos_AR(phi)
}
\arguments{
\item{phi}{Vector of AR coefficients.}
}
\value{
Complex vector of roots.
}
\description{
Computes the roots of the AR characteristic polynomial.
}
