% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btwar_fit.R
\name{print.summary.btwar}
\alias{print.summary.btwar}
\title{Print a BTWAR Model Summary}
\usage{
\method{print}{summary.btwar}(x, ...)
}
\arguments{
\item{x}{Object of class \code{"summary.btwar"}.}

\item{...}{Additional arguments (currently unused).}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print a BTWAR Model Summary
}
