% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btwar_fit.R
\name{residuals.btwar}
\alias{residuals.btwar}
\title{Residuals from a BTWAR Model}
\usage{
\method{residuals}{btwar}(object, ...)
}
\arguments{
\item{object}{Object of class \code{"btwar"}.}

\item{...}{Additional arguments (currently unused).}
}
\value{
Numeric vector of residuals from the training set.
}
\description{
Residuals from a BTWAR Model
}
