% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ar_tools.R
\name{yhat_ar}
\alias{yhat_ar}
\title{Generate AR One-Step-Ahead Predictions}
\usage{
yhat_ar(x, phi)
}
\arguments{
\item{x}{Time series.}

\item{phi}{AR coefficient vector.}
}
\value{
Numeric vector of predictions.
}
\description{
Computes recursive one-step-ahead predictions
from an autoregressive model.
}
