% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ar_tools.R
\name{yhat_arma}
\alias{yhat_arma}
\title{Generate ARMA Predictions}
\usage{
yhat_arma(x, ar, ma)
}
\arguments{
\item{x}{Time series.}

\item{ar}{AR coefficients.}

\item{ma}{MA coefficients.}
}
\value{
Numeric vector of predictions.
}
\description{
Computes recursive ARMA predictions.
}
